Change how "canceling" works
Before, revoking a token or setting a new auth token would kill the
token retrieval processes. Now, when that happens, any in-flight token
fetches will stop what they are doing and try to get a new token.

This is useful to ensure that the freshest token is retrieved without
the user needing to know that the token error was a cancellation, so
he/she can try again.